chore(tests): moving tests around, splitting & exporting for unit testing
