+++ test/unit/activation/find-process-activatable.spec.ts
@@ -1,0 +1,9 @@
+import { processActivatable, findActivatable } from '../../../src/activation';
+
+describe('Activation -- findActivatable()', function _1_findActivatable__Tests() {
+
+});
+
+describe('Activation -- processActivatable()', function _1_processActivatable__Tests() {
+
+});
