+++ test/unit/activation/find-process-deactivatable.spec.ts
@@ -1,0 +1,9 @@
+import { findDeactivatable, processDeactivatable } from '../../../src/activation';
+
+describe('Activation -- processDeactivatable', () => {
+
+});
+
+describe('Activation -- findDeactivatable()', () => {
+  // findDeactivatable
+});
