+++ test/unit/route-loading/load-route.spec.ts
@@ -1,5 +1,5 @@
 import { loadRoute } from '../../../src/route-loading';
 
-fdescribe('RouteLoading -- loadRoute()', function loadRoute__Tests() {
+describe('RouteLoading -- loadRoute()', function loadRoute__Tests() {
 
 });
